api/requirements.txt
Django==3.2.6
psycopg2==2.9.1
djangorestframework==3.12.4
django-filter==2.4.0
django-cors-headers==3.8.0
requests==2.26.0
# Develop
autopep8
flake8